refactor(패키지 분리) : utility 및 service 패키지로 이동
 - 입력 클래스는 view 패키지로 이동
 - RandomNumber 클래스는 utility로 이동
 - CarRacingProgram 클래스는 service 패키지로 이동
 - CarRacingProgram 리스트 초기화 구문 생성자에서 변수 선언 시 할당으로 변경
