+++ src/main/java/carracing/MainApplication.java
@@ -1,6 +1,6 @@
 package carracing;
 
-import carracing.domain.CarRacingProgram;
+import carracing.service.CarRacingProgram;
 
 public class MainApplication {
