+++ src/main/java/carracing/utility/RandomNumber.java
@@ -1,4 +1,4 @@
-package carracing.domain;
+package carracing.utility;
 
 import java.util.Random;
