+++ src/main/java/carracing/view/RacingInfoScanner.java
@@ -1,4 +1,4 @@
-package carracing.domain;
+package carracing.view;
 
 import carracing.view.CarRacingInfoPrint;
